+++ tools/cpp_tools/ranger_ext/Forest.cpp
@@ -134,7 +134,7 @@ void Forest::initCpp(std::string dependent_variable_name, MemoryMode memory_mode
 }
 // #nocov end
 
-void Forest::initCppFromMem(std::string dependent_variable_name, MemoryMode memory_mode, float *x_mem, float *y_mem, int num_rows, int num_cols, uint mtry,
+void Forest::initCppFromMem(std::string dependent_variable_name, MemoryMode memory_mode, float *x_mem, float *y_mem, size_t num_rows, size_t num_cols, uint mtry,
     std::string output_prefix, uint num_trees, std::ostream* verbose_out, uint seed, uint num_threads,
     std::string load_forest_filename, ImportanceMode importance_mode, uint min_node_size,
     std::string split_select_weights_file, const std::vector<std::string>& always_split_variable_names,
@@ -1047,7 +1047,7 @@ std::unique_ptr<Data> Forest::loadDataFromFile(const std::string& data_path) {
 }
 // #nocov end
 
-std::unique_ptr<Data> Forest::loadDataFromMem(float* x_mem, float* y_mem, int num_rows, int num_cols) {
+std::unique_ptr<Data> Forest::loadDataFromMem(float* x_mem, float* y_mem, size_t num_rows, size_t num_cols) {
   if (!x_mem || !y_mem) {
     *this->verbose_out << __FILE__ << ":" << __LINE__ << "Error: null mem ptrs" << std::endl;
     return nullptr;
